FITSBI - backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.13.1
python-dotenv==1.0.0
openai==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
pydantic>=2.5.0
cryptography==41.0.5
--only-binary :all:
